requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
asyncpg==0.29.0
redis==5.0.4
pydantic==2.7.1
pydantic-settings==2.2.1
jinja2==3.1.4
aiofiles==23.2.1
httpx==0.27.0
pyjwt==2.8.0
python-dotenv==1.0.1
